Adding a musical brace to the left-side of the Panel, for aesthetic impact, as a kind of border.
I created a .svg file of a musical brace, using online .svg makers and .svg drawing applications. File is now included in repo.

You can't set the brace as the border image very effectively (restrictions on width). I am instead experimenting with putting the brace into a different, parent Div element, to wrap it around the Panel and make the brace grow/shrink height with the panel as needed.

I have also experimented with having the brace appear as a ::before css property for the Panel, but the brace wasn't visible.