Implemented Brace using panel::before. Needed to crop the original file in order to remove leading and trailing whitespace, so new file, suffix -cropped.svg, has been uploaded.
Ran into issue with scaling the .svg file. Some sources say it's impossible to scale an svg file by only x or y direction, and I was really struggling with getting the brace to obey min-width / min-height etc. The solution is to go into the <svg> properties in the .svg file, and set preserveAspectRatio = "none".