Shifted color, size, position, width of Brace in panel::before. Moderately satisfied with the outcome (dark gray brace). Rounded out the Panel corners from 20px to 30px to look more similar to brace.
Converted the brace .svg image from background-image to mask-image, so that the brace color can be changed from black. I landed on black anyway, and use an overlay blend mode.

TODO:
May want to add a vertical bar to the right of the brace to continue the visual metaphor of sheet music. May want to also include 5 perpendicular lines on top and bottom to imitate staves, fading behind chords. May be too visually busy, and visual metaphor breaks with Chords shaped like autoharp keys.

Alternatively, may want to restyle the left edge of the panel so that no part of the panel passes further left of the brace. (Trimmed / intersection). Not sure how to do this yet.