Created getInputChords and onInputChordsChanged helper functions.
getInputChords scrapes the chord list from queryselectorall '.chord span'. Since the chords are on the dom in order, they will return in order on a selector query.

onInputChordsChanged calls getinputchords, and then passess the inputChords to calculateResultingChords. Then, for every chord returned, it creates a chord and appends it to chordgroup2.

TODO: If there are multiple sets, create a new chordgroup(?) for each set.